server/pyproject.toml
[tool.poetry]
name = "microerp"
version = "0.1.0"
description = ""
authors = ["Gabriel Aguiar"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
drf-yasg = "^1.21.7"
python-dotenv = "^1.0.1"
django-allauth = "^0.63.3"
django-cors-headers = "^4.3.1"
pillow = "^10.3.0"
setuptools = "^70.0.0"
psycopg = {extras = ["binary"], version = "^3.1.19"}
dj-database-url = "^2.2.0"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
gunicorn = "^22.0.0"
uvicorn = "^0.30.1"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
pylint = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
